Move engine::global, engine::utilities, log -> global
