+++ godot-codegen/src/special_cases/special_cases.rs
@@ -37,7 +37,7 @@ pub fn is_class_method_deleted(class_name: &TyName, method: &JsonClassMethod, ct
     }
     
     match (class_name.godot_ty.as_str(), method.name.as_str()) {
-        // Already covered by manual APIs
+        // Already covered by manual APIs.
         //| ("Object", "to_string")
         | ("Object", "get_instance_id")
         
@@ -280,8 +280,19 @@ pub fn is_builtin_type_scalar(name: &str) -> bool {
     name.chars().next().unwrap().is_ascii_lowercase()
 }
 
+#[rustfmt::skip]
 pub fn is_utility_function_deleted(function: &JsonUtilityFunction, ctx: &mut Context) -> bool {
-    codegen_special_cases::is_utility_function_excluded(function, ctx)
+    let hardcoded = match function.name.as_str() {
+        // We have prints manually defined, with (&[Variant]) instead of (Variant, &[Variant]) signature.
+        | "print"
+        | "print_rich"
+        // Should we map str, printt, prints, printerr, printraw manually as well?
+
+        => true, _ => false
+    };
+
+
+    hardcoded || codegen_special_cases::is_utility_function_excluded(function, ctx)
 }
 
 pub fn maybe_rename_class_method<'m>(class_name: &TyName, godot_method_name: &'m str) -> &'m str {
