+++ godot-core/src/global/mod.rs
@@ -1,0 +1,17 @@
+/*
+ * Copyright (c) godot-rust; Bromeon and contributors.
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+
+//! Godot global enums, constants and utility functions.
+//!
+//! See also [Godot docs for `@GlobalScope`](https://docs.godotengine.org/en/stable/classes/class_@globalscope.html#methods).
+
+pub(crate) mod print;
+
+pub use crate::gen::central::global_enums::*;
+pub use crate::gen::utilities::*;
+
+pub use print::*;
