+++ godot/src/prelude.rs
@@ -15,12 +15,17 @@ pub use super::builtin::math::FloatExt as _;
 pub use super::builtin::meta::{FromGodot, ToGodot};
 
 pub use super::engine::{
-    load, try_load, utilities, AudioStreamPlayer, Camera2D, Camera3D, GFile, IAudioStreamPlayer,
-    ICamera2D, ICamera3D, INode, INode2D, INode3D, IObject, IPackedScene, IRefCounted, IResource,
-    ISceneTree, Input, Node, Node2D, Node3D, Object, PackedScene, RefCounted, Resource, SceneTree,
+    load, try_load, AudioStreamPlayer, Camera2D, Camera3D, GFile, IAudioStreamPlayer, ICamera2D,
+    ICamera3D, INode, INode2D, INode3D, IObject, IPackedScene, IRefCounted, IResource, ISceneTree,
+    Input, Node, Node2D, Node3D, Object, PackedScene, RefCounted, Resource, SceneTree,
 };
+pub use super::global::{
+    godot_error, godot_print, godot_print_rich, godot_script_error, godot_warn, EulerOrder, Side,
+    VariantOperator,
+};
+/* ProjectionEye etc */
+
 pub use super::init::{gdextension, ExtensionLibrary, InitLevel};
-pub use super::log::*;
 pub use super::obj::{Base, Gd, GdMut, GdRef, GodotClass, Inherits, InstanceId, OnReady};
 
 // Make trait methods available.
